WiP [ansible] Direct image download from Glance
as Ansible deploy driver is communicating with the node over encrypted
transport (SSH) it is not that great security risk to pass in a token
that can be used to download the image from the Glance API directly.

For more security, use a dedicated domain,
create temporary users in this domain and create a single-use impersonating
keystone trust between user from request context and temporary user.
Then create a token with this trust, and pass it to Ansible to download
glance image directly from glance API.
After Ansible finished (successfully or not) delete the temporary user,
which also deletes all the trusts issued for it.

Change-Id: I280b5063694448bfb26e43f04022317a5dae825b
Depends-On: I803aa28fcd52e5244c473c90cafb80fb5893c7ee