[ansible] PoC Direct image download from Glance
as Ansible deploy driver is communicating with the node over encrypted
transport (SSH) it is not that great security risk to pass in a token
that can be used to download the image from the Glance API directly.

OTOH impersonating the ramdisk is still possible by intercepting it
booting over (i)PXE, so leave this patch as PoC for future reference.

Change-Id: I280b5063694448bfb26e43f04022317a5dae825b