[ansible] imporve logging in DevStack
DevStack has peculiar logging configuration, where logging is performed
to stdout of a process running in screen and `tee`d to an actual log
file.

This makes it impossible to discover the log file from ironic.conf only,
and thus previously required setting the log file explicitly in
configuration file for ironic's Ansible callback plugin.
Nevetheless it was not too friendly for developers, as it required
simultaneous monitoring of ir-cond.log file (or a separate configured
ansible log file) instead of just watching the log live in screen as
usual.

Whith this patch, when no log file can be inferred from logging
configuration set in ironic.conf,
and no explicit log file is specified in callback plugin settings file,
callback plugin assumes it is being run in DevStack and
tries to find the pid of the running ironic-conductor process to use
its stdout file descriptor as a log file.

Change-Id: I365a82cd3592a7669599fe84674ca4abab8f576f