WiP Add network interface for neutron + no nova
Primary use case is standalone ironic deployment (no nova) that wants to
use Neutron for IP management when flat shared network is used for
baremetal nodes.

This interface basically creates 'tenant' ports itself instead of
relying on those ports being created by nova.
The final IP of the node is saved as 'ipv4_addresses' field in the
node's instance_info so it can be retrieved later.

Change-Id: I494ecde81048a4b96dcf59d4c377b712572cc2ec