WiP [ansible] Add ansible-deploy vendor passthru
the new passthru method `list_ansible_files` returns the list of files
available to set as various `ansible_deploy_*` driver info fields
together with some descriptions of those scrobbled from files themselves.

It is implemented as node passthru method since generally available
playbooks et al can differ per-conductor (although we really discourage
this) and thus depend on node's conductor affinity.

The parsed files are cached on driver load.
If this method is called with GET, cached version is returned as-is.
If this method is called with POST, cache is refreshed first by
re-reading disk content.
No other HTTP methods or arguments are supported.

The call is syncronous but does not require an exclusive lock on node.

Change-Id: Ic1f078d3f2646adc2c984ff8b55a81763be8eb50