[ansible] Add ansible-deploy vendor passthru
the new passthru method `list_ansible_files` returns the list of files
available to set as various `ansible_deploy_*` driver info fields
together with some descriptions of those scrobbled from files themselves.

It is implemented as node passthru method since generally available
playbooks and clean steps configs can differ per-conductor
(although we really discourage this)
and thus depend on node's conductor affinity.

The only supported HTTP method is POST, with no arguments.
The call is syncronous but does not require an exclusive lock on node.

As classic drivers are about to be deprecated,
this vendor interface is only declared as supported for
ansible-enabled hardware types in `ironic_staging_drivers` package,
which are`staging-ansible-ipmi` and `staging-libvirt`.

Change-Id: Ic1f078d3f2646adc2c984ff8b55a81763be8eb50