Add hardware types
Classic drivers will be deprecated, add hardware types for
staging drivers. "staging-" prefix used for entrypoints.

The DevStack plugin is amended to enable all hardware types and
interfaces to test that they do not break ironic-conductor start when
enabled (as previousely, iboot drivers are still skipped as they require a
client library not available on PyPI).

Change-Id: I26541adc9b99f7a436be0e3f8ef31c0ff64ff5d4