[ansible] skip validating SSL certs
do not validate SSL certs when downloading the image from TempURL.
CI uses self-signed certs which does not get validated that easily.

This would have to be done configurable later,
or devstack plugin will have to be amended to insert those
self-signed certificates to the ramdisk during its rebuild.

Change-Id: Id40f1067361cc32e98973c70fc5cd3d4242308d0
Partial-Bug: #1717858