Include release notes in the documentation
As discussed on IRC (see linked LP bug) our easiest choice
for publishing release notes is to include them as part of
our RTFD-published documentation.

This patch modifies Sphinx config file to understand if docs
are being built from a git checkout, and if so, include
reno-generated release notes in the documentation.

Change-Id: Iaad56d304594a9e3ab6f48fd8fd9829b071ce6a2
Closes-Bug: #1623034