Handle 'timeout' parameter in power methods
The PowerInterface methods set_power_state() and reboot() were enhanced
to take a 'timeout' parameter [1].

In the Queens release [2] , the conductor assumes that all PowerInterfaces
can support this timeout parameter. This patch changes the PowerInterfaces
so that they accept a 'timeout' parameter for reboot() and set_power_state().
The PowerInterfaces log a warning if the parameter is specified when it isn't
supported.

[1] f15d5b9a37260b3876f9dadeb030412e6e1053b2
[2] 9e87cebc12102cbb3ae47366836dcd7c3e439828

Change-Id: I0904958f1a7f981204ba41bfbdc0083182c5622e