WOL driver - include POWER_OFF in supported power states
The WakeOnLanPower class is designed to POWER_ON using WOL and log
REBOOT/POWER_OFF operations as they require manual intervention.

However the POWER_OFF state was mistakenly omitted from
get_supported_power_states, which results in an error instead of a log entry.
Introspection on tripleo fails as a result of this.

Change-Id: I34ee6da8600d6e6107f00732bf1c34c5b745b333