Add noop management to iBOOT and WOL hardware types
The classic drivers had management=None, so no management actions
were tried. The hardware types have management=fake, which does not
support local boot. This ends up in a serious feature regression.

The intended behavior is to use static boot management, which is what
the noop management is about, so use it.

Change-Id: Id46e2747facabda922e1544fa59e90fb31a04a62